import修正、エラーハンドリング追加、引数や返り値を修正
